unify slider widgets
