+++ lib/features/family/features/giving_flow/widgets/slider_widget.dart
@@ -4,9 +4,9 @@ import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:givt_app/core/enums/enums.dart';
 import 'package:givt_app/features/family/features/giving_flow/create_transaction/cubit/create_transaction_cubit.dart';
 import 'package:givt_app/features/family/features/scan_nfc/cubit/scan_nfc_cubit.dart';
+import 'package:givt_app/features/family/shared/design/components/input/fun_slider.dart';
 import 'package:givt_app/features/family/shared/widgets/loading/custom_progress_indicator.dart';
 import 'package:givt_app/features/family/shared/widgets/texts/shared_texts.dart';
-import 'package:givt_app/features/family/utils/family_app_theme.dart';
 import 'package:givt_app/utils/utils.dart';
 
 class SliderWidget extends StatelessWidget {
@@ -32,19 +32,7 @@ class SliderWidget extends StatelessWidget {
           ),
         ),
         SliderTheme(
-          data: SliderTheme.of(context).copyWith(
-            trackHeight: 7,
-            activeTrackColor: Theme.of(context).colorScheme.onInverseSurface,
-            thumbShape: const SliderWidgetThumb(thumbRadius: 17),
-            inactiveTrackColor:
-                Theme.of(context).colorScheme.surfaceContainerHighest,
-            activeTickMarkColor: Theme.of(context).colorScheme.onInverseSurface,
-            inactiveTickMarkColor:
-                Theme.of(context).colorScheme.surfaceContainerHighest,
-            valueIndicatorColor: Colors.white,
-            thumbColor: Theme.of(context).colorScheme.onInverseSurface,
-            disabledThumbColor: FamilyAppTheme.secondary30,
-          ),
+          data: FunSliderTheme.getSliderTheme(context),
           child: Slider(
             value: currentAmount,
             max: maxAmount,
@@ -82,51 +70,5 @@ class SliderWidget extends StatelessWidget {
         ),
       ],
     );
-  }
-}
-
-class SliderWidgetThumb extends SliderComponentShape {
-  const SliderWidgetThumb({
-    required this.thumbRadius,
-  });
-  final double thumbRadius;
-
-  @override
-  Size getPreferredSize(bool isEnabled, bool isDiscrete) {
-    return Size.fromRadius(thumbRadius);
-  }
-
-  @override
-  void paint(
-    PaintingContext context,
-    Offset center, {
-    required Animation<double> activationAnimation,
-    required Animation<double> enableAnimation,
-    required bool isDiscrete,
-    required TextPainter labelPainter,
-    required RenderBox parentBox,
-    required SliderThemeData sliderTheme,
-    required TextDirection textDirection,
-    required double value,
-    required double textScaleFactor,
-    required Size sizeWithOverflow,
-  }) {
-    // Draw circle
-    final circleBluePaint = Paint()
-      ..color = sliderTheme.thumbColor!
-      ..style = PaintingStyle.fill;
-
-    final circleDarkPaint = Paint()
-      ..color = FamilyAppTheme.secondary30
-      ..style = PaintingStyle.fill;
-
-    // Draw shadow
-    context.canvas
-        .drawCircle(center + const Offset(0, 2), thumbRadius, circleDarkPaint);
-
-    context.canvas.drawCircle(center, thumbRadius, circleBluePaint);
-
-    // Inner circle
-    context.canvas.drawCircle(center, 6, circleDarkPaint);
   }
 }
