+++ lib/features/family/shared/design/components/input/input.dart
@@ -1,1 +1,2 @@
 export 'fun_counter.dart';
+export 'fun_slider.dart';
